+++ blog/src/main/java/com/javaproject/blog/controllers/AnswerController.java
@@ -1,21 +1,22 @@
 package com.javaproject.blog.controllers;
-
-import com.javaproject.blog.repo.QuestionRepository;
-import com.javaproject.blog.models.Question;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 
-@Controller
-public class QuestionController {
-    @Autowired
-    private QuestionRepository questionRepository;
+import com.javaproject.blog.models.Answer;
+import com.javaproject.blog.repo.AnswerRepository;
 
-    @GetMapping("/questions")
+@Controller
+public class AnswerController {
+
+    @Autowired
+    private AnswerRepository answerRepository;
+
+    @GetMapping("/answers")
 	public String session(Model model) {
-		Iterable<Question> questions = questionRepository.findAll();
-		model.addAttribute("questions", questions);
+		Iterable<Answer> answers = answerRepository.findAll();
+		model.addAttribute("answers", answers);
 		return "session";
 	}
 }
