+++ blog/src/main/java/com/javaproject/blog/controllers/UserControllers.java
@@ -1,20 +1,19 @@
 package com.javaproject.blog.controllers;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 
-import com.javaproject.blog.repo.AnswerRepository;
+import com.javaproject.blog.repo.UserRepository;
 
 @Controller
-public class AnswersController {
-
+public class UserControllers {
     @Autowired
-    private AnswerRepository answersRepository;
-
-    @GetMapping("/")
-	public String session(Model model) {
-		model.addAttribute("title", "Авторизация");
-		return "";
+    private UserRepository userRepository;
+    @GetMapping("/registration")
+	public String home(Model model) {
+		model.addAttribute("title", "Главная страница");
+		return "registration";
 	}
 }
