+++ blog/src/main/resources/application.properties
@@ -1,7 +1,7 @@
 server.port = 3000
 
 spring.jpa.hibernate.ddl-auto=update
-spring.datasource.url=jdbc:mysql://${MYSQL_HOST:localhost}:3306/db_example
+spring.datasource.url=jdbc:mysql://${MYSQL_HOST:localhost}:3306/spring_web_blog
 spring.datasource.username=root
 spring.datasource.password=
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
