+++ blog/src/main/resources/templates/registration.html
@@ -14,12 +14,14 @@
     <div class='article'>
       <div class="card-body bg-dark text-white">
         <form action="/users" accept-charset="UTF-8" data-remote="true" method="post"><input name="utf8" type="hidden" value="&#x2713;" /><input type="hidden" name="authenticity_token" value="/RhOquRwDBhYmO12ZgTX0W/b5m1bj6aqN3WCZUSLkj03M5f7ZoLDLgeg2PyvM4vXR3YRxVRmdCJ/Nwud81zq+A==" />
+          
           <div class="mb-3 row">
             <label class="col-sm-2 col-form-label" for="user_E-Mail">E-mail</label><br/>
             <div class="col-sm-10">
               <input required="required" minlegth="7" placeholder="example@email.com" class="form-control" type="email" name="user[email]" id="user_email" />
             </div>
           </div>
+
           <div class="mb-3 row"> 
             <label class="col-sm-2 col-form-label" for="user_Ваше Имя">Ваше Имя</label><br/>
             <div class="col-sm-10">
